requirements.txt
Flask
python-dotenv
psycopg2-binary
flask-cors
supabase
gunicorn